fixed: minor signed/unsigned compile warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@10917 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/FileSystem/CacheStrategy.cpp
@@ -147,7 +147,7 @@ int CSimpleFileCache::ReadFromCache(char *pBuffer, size_t iMaxSize)
 	}
 
 	if (iMaxSize > iAvailable)
-		iMaxSize = iAvailable;
+		iMaxSize = (size_t)iAvailable;
 
 	DWORD iRead = 0;
 	if (!ReadFile(m_hCacheFileRead, pBuffer, iMaxSize, &iRead, NULL)) {
@@ -203,7 +203,7 @@ __int64 CSimpleFileCache::Seek(__int64 iFilePosition, int iWhence) {
 	}
 
 	__int64 nDiff = iTarget - m_nWritePosition;
-	if ( nDiff > 500000 || (nDiff > 0 && WaitForData(nDiff, 5000) == CACHE_RC_TIMEOUT)  ) {		
+	if ( nDiff > 500000 || (nDiff > 0 && WaitForData((unsigned int)nDiff, 5000) == CACHE_RC_TIMEOUT)  ) {		
 		CLog::Log(LOGWARNING,"%s - attempt to seek pass read data (seek to %lld. max: %lld. reset read pointer. (%lld:%d)", __FUNCTION__, iTarget, m_nWritePosition, iFilePosition, iWhence);
 		return  CACHE_RC_ERROR;
 	}
